# backend/swagger.yaml
openapi: 3.0.3
info:
  title: MyActivities API
  description: |
    API REST de la plateforme **MyActivities** — découverte, planification et évaluation d'activités locales.

    ## Authentification
    L'API utilise des **JWT Bearer tokens** (access token 15 min + refresh token 7 jours).
    Les endpoints protégés nécessitent le header : `Authorization: Bearer <access_token>`

    ## Rôles utilisateurs
    | Rôle | Description |
    |------|-------------|
    | `visitor` | Non connecté — lecture seule |
    | `member` | Compte créé — planning, avis |
    | `organizer` | Référence ses activités |
    | `admin` | Modération et administration |

    ## Règles métier importantes
    - **1 avis par utilisateur par activité** (enforced côté backend)
    - **Toute nouvelle activité requiert une validation admin** avant publication
    - **Brute force** : 5 tentatives échouées → blocage 15 min
    - **Géolocalisation** : consentement explicite requis (RGPD Art. 6.1.a)
    - **Suppression de compte** : efface toutes les données personnelles (RGPD Art. 17)

    ## Codes d'erreur communs
    | Code | Signification |
    |------|---------------|
    | 400 | Données invalides |
    | 401 | Token manquant ou expiré |
    | 403 | Accès refusé (rôle insuffisant) |
    | 404 | Ressource introuvable |
    | 409 | Conflit (ex : email déjà utilisé) |
    | 429 | Trop de requêtes (rate limiting) |

  version: 1.0.0
  contact:
    name: Charles Deschamps
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT (expiration 15 min)

  # ─────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 3

    # ── Errors ──────────────────────────────────
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Le champ email est invalide.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── Auth ────────────────────────────────────
...
    ActivityCategory:
      type: string
      enum: [sport, culture, gastronomie, nature, bien_etre, art, musique, autre]
...
    ActivitySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Escalade en salle — ClimbUp Lyon
        category:
          $ref: '#/components/schemas/ActivityCategory'
        distance:
          type: number
          format: float
          description: Distance en km depuis la position de l'utilisateur
          example: 1.4
        avgRating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.2
        reviewCount:
          type: integer
          example: 37
        priceMin:
          type: number
          format: float
          example: 8.0
        priceMax:
          type: number
          format: float
          example: 15.0
        coverImage:
          type: string
          format: uri
          nullable: true
        address:
          type: string
          example: 12 Rue de la République, 69001 Lyon
        latitude:
          type: number
          format: float
          example: 45.7640
        longitude:
          type: number
          format: float
          example: 4.8357
...
    PlanningEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        activity:
          $ref: '#/components/schemas/ActivitySummary'
        scheduledAt:
          type: string
          format: date-time
          example: "2026-07-14T14:00:00Z"
        reminderOffset:
          type: integer
          description: Délai de rappel en minutes avant l'activité
          example: 60
          nullable: true
        createdAt:
          type: string
          format: date-time
    AddToPlanningRequest:
      type: object
      required: [activityId, scheduledAt]
      properties:
        activityId:
          type: string
          format: uuid
        scheduledAt:
          type: string
          format: date-time
          description: La date ne peut pas être dans le passé
          example: "2026-07-14T14:00:00Z"
        reminderOffset:
          type: integer
          description: "Délai de rappel en minutes (ex: 60 = 1h avant)"
          example: 60
          nullable: true

    # ── Reviews ──────────────────────────────────
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        content:
          type: string
          example: Bonjour, l'activité est-elle accessible en poussette ?
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
        sender:
          type: object
          properties:
            id:
              type: string
              format: uuid
            pseudo:
              type: string
            avatarUrl:
              type: string
              format: uri
              nullable: true
    ConversationParticipant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [admin, member]
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [direct, group]
        title:
          type: string
          nullable: true
          description: Nom du groupe (null pour une conversation directe).
        otherParticipant:
          type: object
          nullable: true
          description: L'autre membre d'un direct (null pour un groupe).
          properties:
            id:
              type: string
              format: uuid
            pseudo:
              type: string
            avatarUrl:
              type: string
              format: uri
              nullable: true
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ConversationParticipant'
        lastMessage:
          type: object
          nullable: true
          properties:
            content:
              type: string
            createdAt:
              type: string
              format: date-time
            senderId:
              type: string
              format: uuid
        unreadCount:
          type: integer
          example: 3
        updatedAt:
          type: string
          format: date-time
...
    SendMessageRequest:
      type: object
      required: [recipientId, content]
      properties:
        recipientId:
          type: string
          format: uuid
        content:
          type: string
          minLength: 1
          maxLength: 2000
          example: Bonjour, l'activité est-elle accessible en poussette ?
    SendToConversationRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
    CreateGroupRequest:
      type: object
      required: [title, memberIds]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
          example: Rando dimanche
        memberIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          maxItems: 50
    RenameGroupRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
    AddMemberRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          format: uuid
...
  parameters:
...
    planningId:
      name: planningId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ─────────────────────────────────────────────
  # RESPONSES
  # ─────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Token manquant ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token JWT manquant ou expiré.
    Forbidden:
      description: Rôle insuffisant pour cette action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: Accès réservé aux administrateurs.
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Ressource introuvable.
...
paths:
...
  /planning:
    get:
      tags: [Planning]
      summary: Obtenir son planning
      description: Retourne les activités planifiées par l'utilisateur connecté, triées par date croissante.
      operationId: getPlanning
      security:
        - BearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Filtrer à partir de cette date (ISO 8601)
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: Filtrer jusqu'à cette date (ISO 8601)
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Planning de l'utilisateur
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlanningEntry'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Planning]
      summary: Ajouter une activité au planning
      description: La date scheduledAt ne peut pas être dans le passé.
      operationId: addToPlanning
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToPlanningRequest'
      responses:
        '201':
          description: Activité ajoutée au planning
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningEntry'
        '400':
          description: Date invalide ou dans le passé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Activité introuvable ou non publiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /planning/{planningId}:
    delete:
      tags: [Planning]
      summary: Supprimer une entrée du planning
      operationId: removePlanning
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/planningId'
      responses:
        '204':
          description: Entrée supprimée du planning
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: L'entrée n'appartient pas à l'utilisateur connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════
  # REVIEWS
  # ══════════════════════════════
...
  /messages:
    post:
      tags: [Messaging]
      summary: Envoyer un message
      description: |
        Envoie un message à un autre utilisateur. La conversation est créée à la volée
        si elle n'existe pas (une seule conversation par paire d'utilisateurs).
        Équivalent REST de l'événement Socket.IO `message:send` : déclenche aussi la
        diffusion temps réel `message:new` vers les deux participants.
      operationId: sendMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message envoyé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Destinataire invalide (ex. envoi à soi-même)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_RECIPIENT
                message: Impossible de s'envoyer un message à soi-même.
  /messages/conversations:
    get:
      tags: [Messaging]
      summary: Lister ses conversations
      description: |
        Conversations de l'utilisateur connecté, triées par activité récente,
        avec le dernier message et le nombre de messages non lus.
      operationId: listConversations
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Liste paginée des conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/conversations/{conversationId}:
    get:
      tags: [Messaging]
      summary: Historique d'une conversation
      description: Messages paginés d'une conversation, du plus récent au plus ancien.
      operationId: getMessages
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Liste paginée des messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Messaging]
      summary: Envoyer un message dans une conversation existante
      description: |
        Envoie un message vers une conversation existante (directe ou groupe).
        L'utilisateur doit en être membre. Équivalent REST de `message:send-group` :
        diffuse `message:new` à tous les participants.
      operationId: sendToConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendToConversationRequest'
      responses:
        '201':
          description: Message envoyé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/conversations/{conversationId}/read:
    post:
      tags: [Messaging]
      summary: Marquer une conversation comme lue
      description: |
        Marque comme lus tous les messages reçus non lus de la conversation, puis
        notifie l'autre participant en temps réel (`conversation:read`).
      operationId: markConversationRead
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Conversation marquée lue
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId:
                    type: string
                    format: uuid
                  updated:
                    type: integer
                    description: Nombre de messages passés à l'état lu
                    example: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════
  #  GROUPES
  # ══════════════════════════════
  /messages/groups:
    post:
      tags: [Messaging]
      summary: Créer un groupe
      description: |
        Crée une conversation de groupe. Le créateur en devient administrateur ;
        tous les membres invités doivent être ses amis. Notifie les membres
        (`conversation:new`).
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Groupe créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Membres invalides (non amis, ou aucun autre membre)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/groups/{conversationId}:
    patch:
      tags: [Messaging]
      summary: Renommer un groupe
      description: Renomme un groupe (administrateur requis). Notifie les participants (`conversation:updated`).
      operationId: renameGroup
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameGroupRequest'
      responses:
        '200':
          description: Groupe renommé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/groups/{conversationId}/members:
    post:
      tags: [Messaging]
      summary: Ajouter un membre à un groupe
      description: |
        Ajoute un membre à un groupe (administrateur requis ; le nouveau membre
        doit être ami de l'administrateur). Notifie le nouveau (`conversation:new`)
        et les autres (`conversation:updated`).
      operationId: addGroupMember
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Membre ajouté
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Déjà membre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Utilisateur non ami
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/groups/{conversationId}/members/{userId}:
    delete:
      tags: [Messaging]
      summary: Retirer un membre ou quitter un groupe
      description: |
        Si `userId` est l'utilisateur courant, il quitte le groupe (un membre
        restant est promu administrateur si besoin). Sinon, l'administrateur retire
        le membre visé. Notifie les participants (`conversation:updated`).
      operationId: removeGroupMember
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Membre retiré / groupe quitté
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════
  #  AMIS
  # ══════════════════════════════
...